+++ aasrp/management/commands/aasrp_migrate_srp_data.py
@@ -14,7 +14,6 @@ from eveuniverse.models import EveType
 
 # AA SRP
 from aasrp.helper.character import get_user_for_character
-from aasrp.managers import SrpManager
 from aasrp.models import RequestComment, SrpLink, SrpRequest
 
 
@@ -148,7 +147,7 @@ class Command(BaseCommand):
                                     name=srp_userrequest.srp_ship_name
                                 )
                             except EveType.DoesNotExist:
-                                srp_kill_link = SrpManager.get_kill_id(
+                                srp_kill_link = SrpRequest.objects.get_kill_id(
                                     srp_userrequest_killboard_link
                                 )
 
@@ -156,7 +155,7 @@ class Command(BaseCommand):
                                     ship_type_id,
                                     ship_value,  # pylint: disable=unused-variable
                                     victim_id,  # pylint: disable=unused-variable
-                                ) = SrpManager.get_kill_data(srp_kill_link)
+                                ) = SrpRequest.objects.get_kill_data(srp_kill_link)
 
                                 (
                                     srp_userrequest_ship,